pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tokentaxi"
version = "1.1.4"
description = "Adaptive rate-limit-aware LLM routing. Bring your own clients."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "tokentaxi  contributors" }]
keywords = ["llm", "routing", "openai", "anthropic", "rate-limit", "fallback"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tiktoken>=0.7",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.7",
    "google-genai>=0.1",
    "mistralai>=1.0",
    "cohere>=5.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
redis     = ["redis[asyncio]>=5.0"]
dashboard = ["streamlit>=1.35", "plotly>=5.0"]
cli       = ["typer>=0.12", "rich>=13.0"]
yaml      = ["pyyaml>=6.0"]
all       = ["tokentaxi [redis,dashboard,cli,yaml]"]
dev       = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "pyyaml>=6.0",
]

[project.scripts]
tokentaxi  = "tokentaxi.cli:app"

[project.urls]
"Homepage"    = "https://github.com/your-org/tokentaxi "
"Bug Tracker" = "https://github.com/your-org/tokentaxi/issues"

# ---------------------------------------------------------------------------
# Tool configurations
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version  = "3.11"
strict          = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["tokentaxi"]
